requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.3
python-docx==0.8.11
six==1.16.0
typing-extensions==3.10.0.0
waitress==2.0.0
Werkzeug==2.0.1
zipp==3.5.0
